Delete client

+++ src/routes/MainMenuOutlet/Client/index.tsx
@@ -3,12 +3,15 @@ import SearchField from "../../../components/SearchField";
 import CRUDListItem from "../../../components/CRUDListItem";
 import Button from "../../../components/Button";
 import Toast from "../../../components/Toast";
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { ToastMessage } from "../../../components/Toast/ToastProps";
 import { scrollBarStyles } from "../../../common/constants";
 import { Client as ClientEntity } from "../../../entities/Client";
 import { deleteClient, getClients } from "../../../services/client";
 import ConfirmModal from "../../../components/ConfirmModal";
+import { logInService } from "../../../services/login";
+import { JwtPayload } from "../../../services/login/LoginProps";
+import { AuthContext } from "../../../context/AuthContext";
 
 const containerStyles = 'w-full h-full flex flex-col p-12 gap-8';
 const titleStyles = 'text-4xl font-bold';
@@ -17,7 +20,9 @@ const itemsStyles = 'flex flex-col max-h-96 overflow-y-auto gap-4 p-2 pr-12';
 const firstColumnStyles = 'flex flex-col gap-4';
 
 const Client = () => {
+  const [clientId, setClientId] = useState<number | undefined>(undefined)
   const [clients, setClients] = useState<ClientEntity[]>([]);
+  const { currentUser } = useContext(AuthContext);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [searchField, setSearchField] = useState<string>('');
   const [showConfirmModal, setShowConfirmModal] = useState<boolean>(false);
@@ -42,10 +47,23 @@ const Client = () => {
   const handleDeleteClient = async (clientId: number | undefined) => {
     setIsLoading(true);
     try {
-      await deleteClient({
-        setReturnMessage,
-        client_id: clientId,
+      /** TODO: Change username for a dinamic one. */
+      const response = await logInService({
+        username: currentUser?.username!,
+        password,
+        setLoading: setIsLoading,
+        setReturnMessage
       });
+      const { access_token }: JwtPayload = await response.data;
+      if (access_token) {
+        deleteClient({
+          client_id: clientId,
+          setReturnMessage,
+          access_token,
+        });
+        setReturnMessage({ message: `Cliente deletado com sucesso!`, variation: 'green' })
+        setClients(prev => prev.filter(client => client.id !== clientId))
+      }
     } finally {
       setIsLoading(false);
     }
@@ -70,6 +88,7 @@ const Client = () => {
               setSearchField(event.target.value);
             }}
             required={false}
+            disabled={clients.length === 0}
           />
           <div className={`${itemsStyles} ${scrollBarStyles}`}>
             {clients.map(({ id, name, registration }) => {
@@ -83,7 +102,10 @@ const Client = () => {
                   <CRUDListItem
                     title={name!}
                     description={registration!}
-                    onClickDelete={() => setShowConfirmModal(true)}
+                    onClickDelete={() => {
+                      setClientId(id)
+                      setShowConfirmModal(true)
+                    }}
                     onClickEdit={() => console.log('edit')}
                   />
                 )
@@ -109,6 +131,10 @@ const Client = () => {
       {
         showConfirmModal && <ConfirmModal
           isLoading={false}
+          onConfirm={() => {
+            handleDeleteClient(clientId)
+            setShowConfirmModal(false)
+          }}
           message="Você tem certeza que deseja deletar esse aluno?"
           onTypePassword={setPassword}
           onCancel={() => setShowConfirmModal(false)}
